OPEN - task 1: Further integrate EZ with JSR http://github.com/JavaMoney/javamoney-examples/issues/issue/1

+++ swing/javamoney-ez/money/src/main/java/org/javamoney/examples/ez/money/importexport/DataMerger.java
@@ -56,11 +56,11 @@ DataMerger
 
     if(TransactionHelper.isSplit(category) == true)
     {
-      addSplit(category, trans.getAmount());
+      addSplit(category, trans.getAmount().doubleValue());
     }
     else
     {
-      addCategory(category, trans.getAmount());
+      addCategory(category, trans.getAmount().doubleValue());
     }
   }
